Move items from `smithy-http` to `smithy-types`
