+++ rust-runtime/aws-smithy-http/Cargo.toml
@@ -11,7 +11,7 @@ license = "Apache-2.0"
 repository = "https://github.com/awslabs/smithy-rs"
 
 [features]
-rt-tokio = ["dep:tokio-util", "dep:tokio", "tokio?/rt", "tokio?/fs", "tokio?/io-util", "tokio-util?/io"]
+rt-tokio = ["aws-smithy-types/rt-tokio"]
 event-stream = ["aws-smithy-eventstream"]
 
 [dependencies]
@@ -27,30 +27,20 @@ pin-project-lite = "0.2.9"
 pin-utils = "0.1.0"
 tracing = "0.1"
 
-# We are using hyper for our streaming body implementation, but this is an internal detail.
-hyper = "0.14.26"
-
-# ByteStream internals
+# For our event stream implementation
 futures-core = "0.3.14"
-tokio = { version = "1.23.1", optional = true }
-tokio-util = { version = "0.7", optional = true }
+hyper = "0.14.26"
 
 [dev-dependencies]
 async-stream = "0.3"
 futures-util = { version = "0.3.16", default-features = false }
 hyper = { version = "0.14.26", features = ["stream"] }
-pretty_assertions = "1.3"
 proptest = "1"
 tokio = { version = "1.23.1", features = [
   "macros",
   "rt",
   "rt-multi-thread",
-  "fs",
-  "io-util",
 ] }
-tokio-stream = "0.1.5"
-tempfile = "3.2.0"
-tracing-test = "0.2.1"
 
 [package.metadata.docs.rs]
 all-features = true
