+++ rust-runtime/aws-smithy-types/Cargo.toml
@@ -11,16 +11,29 @@ license = "Apache-2.0"
 repository = "https://github.com/awslabs/smithy-rs"
 
 [features]
+rt-tokio = ["dep:tokio-util", "dep:tokio", "tokio?/rt", "tokio?/fs", "tokio?/io-util", "tokio-util?/io"]
 test-util = []
 serde-serialize = []
 serde-deserialize = []
 
 [dependencies]
 base64-simd = "0.8"
+bytes = "1"
+bytes-utils = "0.1"
+http = "0.2.3"
+http-body = "0.4.4"
+hyper = "0.14.26"
 itoa = "1.0.0"
 num-integer = "0.1.44"
+pin-project-lite = "0.2.9"
+pin-utils = "0.1.0"
 ryu = "1.0.5"
 time = { version = "0.3.4", features = ["parsing"] }
+
+# ByteStream internals
+futures-core = "0.3.14"
+tokio = { version = "1.23.1", optional = true }
+tokio-util = { version = "0.7", optional = true }
 
 [dev-dependencies]
 base64 = "0.13.0"
@@ -31,6 +44,15 @@ proptest = "1"
 rand = "0.8.4"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+tokio = { version = "1.23.1", features = [
+    "macros",
+    "rt",
+    "rt-multi-thread",
+    "fs",
+    "io-util",
+] }
+tokio-stream = "0.1.5"
+tempfile = "3.2.0"
 
 [package.metadata.docs.rs]
 all-features = true
